+++ pkg/cmd/cmd_test.go
@@ -10,12 +10,12 @@ import (
 	"github.com/alexfalkowski/go-health/pkg/server"
 	"github.com/alexfalkowski/go-service/pkg/cmd"
 	"github.com/alexfalkowski/go-service/pkg/config"
-	"github.com/alexfalkowski/go-service/pkg/grpc"
 	"github.com/alexfalkowski/go-service/pkg/health"
-	healthGRPC "github.com/alexfalkowski/go-service/pkg/health/grpc"
-	healthHTTP "github.com/alexfalkowski/go-service/pkg/health/http"
-	pkgHTTP "github.com/alexfalkowski/go-service/pkg/http"
+	healthGRPC "github.com/alexfalkowski/go-service/pkg/health/transport/grpc"
+	healthHTTP "github.com/alexfalkowski/go-service/pkg/health/transport/http"
 	"github.com/alexfalkowski/go-service/pkg/logger"
+	"github.com/alexfalkowski/go-service/pkg/transport/grpc"
+	pkgHTTP "github.com/alexfalkowski/go-service/pkg/transport/http"
 	. "github.com/smartystreets/goconvey/convey"
 	"go.uber.org/fx"
 )
