+++ pkg/health/module.go
@@ -1,8 +1,8 @@
 package health
 
 import (
-	"github.com/alexfalkowski/go-service/pkg/health/grpc"
-	"github.com/alexfalkowski/go-service/pkg/health/http"
+	"github.com/alexfalkowski/go-service/pkg/health/transport/grpc"
+	"github.com/alexfalkowski/go-service/pkg/health/transport/http"
 	"go.uber.org/fx"
 )
