+++ pkg/health/transport/grpc/grpc_test.go
@@ -9,11 +9,11 @@ import (
 	"github.com/alexfalkowski/go-health/pkg/checker"
 	"github.com/alexfalkowski/go-health/pkg/server"
 	"github.com/alexfalkowski/go-service/pkg/config"
-	pkgGRPC "github.com/alexfalkowski/go-service/pkg/grpc"
-	tokenGRPC "github.com/alexfalkowski/go-service/pkg/grpc/security/token"
 	"github.com/alexfalkowski/go-service/pkg/health"
-	healthGRPC "github.com/alexfalkowski/go-service/pkg/health/grpc"
+	healthGRPC "github.com/alexfalkowski/go-service/pkg/health/transport/grpc"
 	"github.com/alexfalkowski/go-service/pkg/logger/zap"
+	pkgGRPC "github.com/alexfalkowski/go-service/pkg/transport/grpc"
+	tokenGRPC "github.com/alexfalkowski/go-service/pkg/transport/grpc/security/token"
 	"github.com/alexfalkowski/go-service/test"
 	. "github.com/smartystreets/goconvey/convey"
 	"go.uber.org/fx/fxtest"
