+++ pkg/health/transport/http/http_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/alexfalkowski/go-health/pkg/server"
 	"github.com/alexfalkowski/go-service/pkg/config"
 	"github.com/alexfalkowski/go-service/pkg/health"
-	healthHTTP "github.com/alexfalkowski/go-service/pkg/health/http"
-	pkgHTTP "github.com/alexfalkowski/go-service/pkg/http"
+	healthHTTP "github.com/alexfalkowski/go-service/pkg/health/transport/http"
 	"github.com/alexfalkowski/go-service/pkg/logger/zap"
+	pkgHTTP "github.com/alexfalkowski/go-service/pkg/transport/http"
 	"github.com/alexfalkowski/go-service/test"
 	. "github.com/smartystreets/goconvey/convey"
 	"go.uber.org/fx/fxtest"
