+++ pkg/transport/http/http_test.go
@@ -11,11 +11,11 @@ import (
 	"testing"
 
 	"github.com/alexfalkowski/go-service/pkg/config"
-	pkgGRPC "github.com/alexfalkowski/go-service/pkg/grpc"
-	tokenGRPC "github.com/alexfalkowski/go-service/pkg/grpc/security/token"
-	pkgHTTP "github.com/alexfalkowski/go-service/pkg/http"
-	tokenHTTP "github.com/alexfalkowski/go-service/pkg/http/security/token"
 	"github.com/alexfalkowski/go-service/pkg/logger/zap"
+	pkgGRPC "github.com/alexfalkowski/go-service/pkg/transport/grpc"
+	tokenGRPC "github.com/alexfalkowski/go-service/pkg/transport/grpc/security/token"
+	pkgHTTP "github.com/alexfalkowski/go-service/pkg/transport/http"
+	tokenHTTP "github.com/alexfalkowski/go-service/pkg/transport/http/security/token"
 	"github.com/alexfalkowski/go-service/test"
 	. "github.com/smartystreets/goconvey/convey"
 	"go.uber.org/fx/fxtest"
